Committing QuickSort3
I ended up restarting my quick sort class twice due to reaching errors that I could not fix. This is the third creation so it is named QuickSort3. I wanted to save my work from my previous creations while I was working on it so I made a new class for each time I tried it.